qa: add another false positive to Psalm baseline
Several tests have ".dist" config files that a setUp process copies to a path omitting the suffix.
That file is then _included_ by the test in order to do assertions.
Psalm was flagging the include as invalid due to the file not existing in a clean setup (which it will not).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>